Dodanie możliwości przywracania habitsów i podglądu wszystkich

+++ Planer/app/src/main/java/com/example/planer/gui/pages/home/HomeFragment.kt
@@ -11,6 +11,7 @@ import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.planer.R
 import com.example.planer.ViewModel.*
@@ -26,6 +27,8 @@ import com.example.planer.scope.UnscrollableLinearLayoutManager
 import kotlinx.android.synthetic.main.dialod_when_title_empty.view.*
 import kotlinx.android.synthetic.main.dialog_habit.view.*
 import kotlinx.android.synthetic.main.fragment_home.view.*
+import kotlinx.coroutines.launch
+import java.lang.Math.abs
 import java.time.LocalDate
 
 
@@ -41,7 +44,6 @@ class HomeFragment : Fragment() {
 
 
     private val settingViewModel: SettingsViewModel by viewModels()
-
 
 
     // lista tasków do recyclerView
@@ -71,7 +73,12 @@ class HomeFragment : Fragment() {
             list,
             listNotes,
             listTypes,
-            { deleteTaskId, deleteNoteId -> userViewModel.deleteTaskAndNoteById(deleteTaskId, deleteNoteId) },
+            { deleteTaskId, deleteNoteId ->
+                userViewModel.deleteTaskAndNoteById(
+                    deleteTaskId,
+                    deleteNoteId
+                )
+            },
             { updateTask, updateNote -> userViewModel.updateTaskAndNote(updateTask, updateNote) }
         )
         rv.adapter = adapter
@@ -83,7 +90,12 @@ class HomeFragment : Fragment() {
             list,
             listNotes,
             listTypes,
-            { deleteTaskId, deleteNoteId -> userViewModel.deleteTaskAndNoteById(deleteTaskId, deleteNoteId) },
+            { deleteTaskId, deleteNoteId ->
+                userViewModel.deleteTaskAndNoteById(
+                    deleteTaskId,
+                    deleteNoteId
+                )
+            },
             { updateTask, updateNote -> userViewModel.updateTaskAndNote(updateTask, updateNote) }
         )
         rv2.adapter = adapter2
@@ -94,7 +106,12 @@ class HomeFragment : Fragment() {
             list,
             listNotes,
             listTypes,
-            { deleteTaskId, deleteNoteId -> userViewModel.deleteTaskAndNoteById(deleteTaskId, deleteNoteId) },
+            { deleteTaskId, deleteNoteId ->
+                userViewModel.deleteTaskAndNoteById(
+                    deleteTaskId,
+                    deleteNoteId
+                )
+            },
             { updateTask, updateNote -> userViewModel.updateTaskAndNote(updateTask, updateNote) }
         )
         rv3.adapter = adapter3
@@ -105,7 +122,12 @@ class HomeFragment : Fragment() {
             list,
             listNotes,
             listTypes,
-            { deleteTaskId, deleteNoteId -> userViewModel.deleteTaskAndNoteById(deleteTaskId, deleteNoteId) },
+            { deleteTaskId, deleteNoteId ->
+                userViewModel.deleteTaskAndNoteById(
+                    deleteTaskId,
+                    deleteNoteId
+                )
+            },
             { updateTask, updateNote -> userViewModel.updateTaskAndNote(updateTask, updateNote) }
         )
         rv4.adapter = adapter4
@@ -116,7 +138,12 @@ class HomeFragment : Fragment() {
             list,
             listNotes,
             listTypes,
-            { deleteTaskId, deleteNoteId -> userViewModel.deleteTaskAndNoteById(deleteTaskId, deleteNoteId) },
+            { deleteTaskId, deleteNoteId ->
+                userViewModel.deleteTaskAndNoteById(
+                    deleteTaskId,
+                    deleteNoteId
+                )
+            },
             { updateTask, updateNote -> userViewModel.updateTaskAndNote(updateTask, updateNote) }
         )
         rv5.adapter = adapter5
@@ -141,7 +168,7 @@ class HomeFragment : Fragment() {
 
                     checkIfEmpty(adapter2.list, adapter3.list, adapter4.list, adapter5.list, view)
 
-                    hideRestOftasks(0, view)
+                    hideRestOftasks(view)
                 }
             }
         }
@@ -168,55 +195,26 @@ class HomeFragment : Fragment() {
 
 
 
-        //hideRestOftasks(0,  view)
-
-
-
-
-
-//        view.today_title.setOnClickListener {
-//            if (view.today_task_list.isVisible) view.today_task_list.visibility = View.GONE
-//            else view.today_task_list.visibility = View.VISIBLE
-//        }
-//        view.tomorrow_title.setOnClickListener {
-//            if (view.tomorrow_task_list.isVisible) view.tomorrow_task_list.visibility = View.GONE
-//            else view.tomorrow_task_list.visibility = View.VISIBLE
-//        }
-//        view.week_title.setOnClickListener {
-//            if (view.week_task_list.isVisible) view.week_task_list.visibility = View.GONE
-//            else view.week_task_list.visibility = View.VISIBLE
-//        }
-//        view.month_title.setOnClickListener {
-//            if (view.month_task_list.isVisible) view.month_task_list.visibility = View.GONE
-//            else view.month_task_list.visibility = View.VISIBLE
-//        }
-//        view.rest_title.setOnClickListener {
-//            if (view.rest_task_list.isVisible) view.rest_task_list.visibility = View.GONE
-//            else view.rest_task_list.visibility = View.VISIBLE
-//        }
-
         view.today_title.setOnClickListener {
             if (view.today_task_list.isVisible) view.today_task_list.visibility = View.GONE
-            else hideRestOftasks(0,  view)
-
+            else view.today_task_list.visibility = View.VISIBLE
         }
         view.tomorrow_title.setOnClickListener {
             if (view.tomorrow_task_list.isVisible) view.tomorrow_task_list.visibility = View.GONE
-            else hideRestOftasks(1,  view)
+            else view.tomorrow_task_list.visibility = View.VISIBLE
         }
         view.week_title.setOnClickListener {
             if (view.week_task_list.isVisible) view.week_task_list.visibility = View.GONE
-            else hideRestOftasks(2,  view)
+            else view.week_task_list.visibility = View.VISIBLE
         }
         view.month_title.setOnClickListener {
             if (view.month_task_list.isVisible) view.month_task_list.visibility = View.GONE
-            else hideRestOftasks(3,  view)
+            else view.month_task_list.visibility = View.VISIBLE
         }
         view.rest_title.setOnClickListener {
             if (view.rest_task_list.isVisible) view.rest_task_list.visibility = View.GONE
-            else hideRestOftasks(4,  view)
+            else view.rest_task_list.visibility = View.VISIBLE
         }
-
 
 
         // HABITS
@@ -234,28 +232,46 @@ class HomeFragment : Fragment() {
         val rvh = view.habits_list
         val adapterhh = AdapterHabits(
             listHab,
-            {hab -> if(delete_clicked == 1){
-                habitViewModel.deleteHabit(hab)
-            }},
-            {hab -> if(delete_clicked == 0){
-                habitViewModel.updateHabit(Habits(hab.id, hab.name, 0))
-            }}
+            { hab ->
+                if (delete_clicked == 1) {
+                    habitViewModel.deleteHabit(hab)
+                }
+            },
+            { hab ->
+                if (delete_clicked == 0) {
+                    habitViewModel.updateHabit(Habits(hab.id, hab.name, abs(hab.isActive - 1)))
+                }
+            }
         )
 
         rvh.adapter = adapterhh
         rvh.layoutManager = LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false)
 
 
+        var allHabits = mutableListOf<Habits>()
+        var activeHabits = mutableListOf<Habits>()
+
+
         //habitViewModel = ViewModelProvider(this)[HabitViewModel::class.java]
         habitViewModel.readAllData.observe(viewLifecycleOwner) {
             //val io = IO()
             //adapterhh.updateList(io.filterHabits(requireContext(), it.toMutableList()))
-            adapterhh.updateList(it.toMutableList())
+
+            allHabits = mutableListOf()
+            activeHabits = mutableListOf()
+
+            for (habit in it) {
+                if (habit.isActive == 1) {
+                    activeHabits.add(habit)
+                }
+            }
+
+            adapterhh.updateList(activeHabits)
+            allHabits = it.toMutableList()
         }
 
 
         //Toast.makeText(requireContext(), listHab.size.toString(), Toast.LENGTH_SHORT).show()
-
 
 
         view.habits_add.setOnClickListener {
@@ -268,8 +284,7 @@ class HomeFragment : Fragment() {
 
             dialogView.btn_create.setOnClickListener {
 
-                if(dialogView.habit_title.text.toString().replace(" ", "") == "")
-                {
+                if (dialogView.habit_title.text.toString().replace(" ", "") == "") {
                     val builder = androidx.appcompat.app.AlertDialog.Builder(requireContext())
                     val inflater = LayoutInflater.from(requireContext())
                     val dialogView = inflater.inflate(R.layout.dialod_when_title_empty, null)
@@ -280,9 +295,7 @@ class HomeFragment : Fragment() {
                     dialogView.btn_ok.setOnClickListener {
                         alertDialog.cancel()
                     }
-                }
-                else
-                {
+                } else {
                     habitViewModel.addHabit(
                         Habits(
                             name = dialogView.habit_title.text.toString(),
@@ -298,16 +311,57 @@ class HomeFragment : Fragment() {
 
         view.habits_delete.setOnClickListener {
             if (delete_clicked == 0) {
-                view.habits_delete.backgroundTintList = ColorStateList.valueOf(ContextCompat.getColor(requireContext(), R.color.pr1_delete_red))
+                view.habits_delete.backgroundTintList = ColorStateList.valueOf(
+                    ContextCompat.getColor(
+                        requireContext(),
+                        R.color.pr1_delete_red
+                    )
+                )
                 //view.habits_delete.setBackgroundColor(Color.parseColor("#33832508"))
                 delete_clicked = 1
             } else {
-                view.habits_delete.backgroundTintList = ColorStateList.valueOf(ContextCompat.getColor(requireContext(), R.color.pr1_beige_background2))
+                view.habits_delete.backgroundTintList = ColorStateList.valueOf(
+                    ContextCompat.getColor(
+                        requireContext(),
+                        R.color.pr1_beige_background2
+                    )
+                )
                 //view.habits_delete.setBackgroundColor(Color.parseColor("#99F3DEBA"))
                 delete_clicked = 0
             }
         }
 
+        var show_all_clicked = 0
+
+        view.habits_show_all.setOnClickListener {
+            if (show_all_clicked == 0)
+            {
+
+                view.habits_show_all.imageTintList = ColorStateList.valueOf(
+                    ContextCompat.getColor(
+                        requireContext(),
+                        R.color.brown_important_urgent_on
+                    )
+                )
+
+                adapterhh.updateList(allHabits)
+
+                show_all_clicked = 1
+            } else {
+                view.habits_show_all.imageTintList = ColorStateList.valueOf(
+                    ContextCompat.getColor(
+                        requireContext(),
+                        R.color.black
+                    )
+                )
+
+                adapterhh.updateList(activeHabits)
+
+                show_all_clicked = 0
+            }
+
+
+        }
 
 
 
@@ -315,136 +369,68 @@ class HomeFragment : Fragment() {
 
     }
 
-    private fun hideRestOftasks(titleClicked: Int, view: View?) {
+    private fun hideRestOftasks(view: View?) {
 
-        when(titleClicked)
-        {
-            //today
-            0 -> {
-                if(view?.today_task_list?.visibility == View.GONE)
-                    view.today_task_list?.visibility = View.VISIBLE
-                if(view?.tomorrow_task_list?.visibility == View.VISIBLE)
-                    view.tomorrow_task_list?.visibility = View.GONE
-                if(view?.week_task_list?.visibility == View.VISIBLE)
-                    view.week_task_list?.visibility = View.GONE
-                if(view?.month_task_list?.visibility == View.VISIBLE)
-                    view.month_task_list?.visibility = View.GONE
-                if(view?.rest_task_list?.visibility == View.VISIBLE)
-                    view.rest_task_list?.visibility = View.GONE
-            }
+        if (view?.today_task_list?.visibility == View.GONE)
+            view.today_task_list?.visibility = View.VISIBLE
+        if (view?.tomorrow_task_list?.visibility == View.VISIBLE)
+            view.tomorrow_task_list?.visibility = View.GONE
+        if (view?.week_task_list?.visibility == View.VISIBLE)
+            view.week_task_list?.visibility = View.GONE
+        if (view?.month_task_list?.visibility == View.VISIBLE)
+            view.month_task_list?.visibility = View.GONE
+        if (view?.rest_task_list?.visibility == View.VISIBLE)
+            view.rest_task_list?.visibility = View.GONE
 
-            //tomorrow
-            1 -> {
-                if(view?.today_task_list?.visibility == View.VISIBLE)
-                    view.today_task_list?.visibility = View.GONE
-                    view?.tomorrow_task_list?.visibility = View.VISIBLE
-                if(view?.week_task_list?.visibility == View.VISIBLE)
-                    view.week_task_list?.visibility = View.GONE
-                if(view?.month_task_list?.visibility == View.VISIBLE)
-                    view.month_task_list?.visibility = View.GONE
-                if(view?.rest_task_list?.visibility == View.VISIBLE)
-                    view.rest_task_list?.visibility = View.GONE
-            }
-
-            //week
-            2 -> {
-
-                if(view?.today_task_list?.visibility == View.VISIBLE)
-                    view.today_task_list?.visibility = View.GONE
-                if(view?.tomorrow_task_list?.visibility == View.VISIBLE)
-                    view.tomorrow_task_list?.visibility = View.GONE
-                view?.week_task_list?.visibility = View.VISIBLE
-
-                if(view?.month_task_list?.visibility == View.VISIBLE)
-                    view.month_task_list?.visibility = View.GONE
-                if(view?.rest_task_list?.visibility == View.VISIBLE)
-                    view.rest_task_list?.visibility = View.GONE
-
-            }
-
-            //month
-             3-> {
-                 if(view?.today_task_list?.visibility == View.VISIBLE)
-                     view.today_task_list?.visibility = View.GONE
-                 if(view?.tomorrow_task_list?.visibility == View.VISIBLE)
-                     view.tomorrow_task_list?.visibility = View.GONE
-                 if(view?.week_task_list?.visibility == View.VISIBLE)
-                     view.week_task_list?.visibility = View.GONE
-                 view?.month_task_list?.visibility = View.VISIBLE
-                 if(view?.rest_task_list?.visibility == View.VISIBLE)
-                     view.rest_task_list?.visibility = View.GONE
-            }
-
-            //rest
-            4 -> {
-                if(view?.today_task_list?.visibility == View.VISIBLE)
-                    view.today_task_list?.visibility = View.GONE
-                if(view?.tomorrow_task_list?.visibility == View.VISIBLE)
-                    view.tomorrow_task_list?.visibility = View.GONE
-                if(view?.week_task_list?.visibility == View.VISIBLE)
-                    view.week_task_list?.visibility = View.GONE
-                if(view?.month_task_list?.visibility == View.VISIBLE)
-                    view.month_task_list?.visibility = View.GONE
-                view?.rest_task_list?.visibility = View.VISIBLE
-
-            }
-
-        }
     }
 
-    private fun checkIfEmpty(list2: MutableList<Tasks>, list3: MutableList<Tasks>, list4: MutableList<Tasks>, list5: MutableList<Tasks>, view: View) {
+    private fun checkIfEmpty(
+        list2: MutableList<Tasks>,
+        list3: MutableList<Tasks>,
+        list4: MutableList<Tasks>,
+        list5: MutableList<Tasks>,
+        view: View
+    ) {
 
-        if(list2.size == 0)
-        {
+        if (list2.size == 0) {
             view.tomorrow_task_list.visibility = View.GONE
             view.tomorrow_title.visibility = View.GONE
             view.line2.visibility = View.GONE
-        }
-        else
-        {
+        } else {
             view.tomorrow_task_list.visibility = View.VISIBLE
             view.tomorrow_title.visibility = View.VISIBLE
             view.line2.visibility = View.VISIBLE
         }
 
-        if(list3.size == 0)
-        {
+        if (list3.size == 0) {
             view.week_task_list.visibility = View.GONE
             view.week_title.visibility = View.GONE
             view.line3.visibility = View.GONE
-        }
-        else
-        {
+        } else {
             view.week_task_list.visibility = View.VISIBLE
             view.week_title.visibility = View.VISIBLE
             view.line3.visibility = View.VISIBLE
 
         }
 
-        if(list4.size == 0)
-        {
+        if (list4.size == 0) {
             view.month_task_list.visibility = View.GONE
             view.month_title.visibility = View.GONE
             view.line4.visibility = View.GONE
 
-        }
-        else
-        {
+        } else {
             view.month_task_list.visibility = View.VISIBLE
             view.month_title.visibility = View.VISIBLE
             view.line4.visibility = View.VISIBLE
 
         }
 
-        if(list5.size == 0)
-        {
+        if (list5.size == 0) {
             view.rest_task_list.visibility = View.GONE
             view.rest_title.visibility = View.GONE
             view.line5.visibility = View.GONE
 
-        }
-        else
-        {
+        } else {
             view.rest_task_list.visibility = View.VISIBLE
             view.rest_title.visibility = View.VISIBLE
             view.line5.visibility = View.VISIBLE
@@ -455,7 +441,6 @@ class HomeFragment : Fragment() {
 
 
 }
-
 
 
 // może się przydać, więc nie usuwam
